Added modules, replaced bass.py with bass_functions.py, and added main analysis pipline as a method of BASS_dataset in BASS.py
